refactor: Reorganize code.
