+++ app/utils/theme/types.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+
+export const colorSchemeSchema = z.enum(['light', 'dark']);
+export type ColorScheme = z.infer<typeof colorSchemeSchema>;
